requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
exceptiongroup==1.2.0
Flask==3.0.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.86
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.9.0
uber-rides==0.6.0
urllib3==2.2.0
Werkzeug==3.0.1
yarl==1.9.4
